(vss version: 4, vss user: Kay)

+++ SuRun/lsa_laar.cpp
@@ -62,7 +62,7 @@ public:
   }
   void Set( char *s )
   {
-    int l=strlen( s ) * sizeof(wchar_t);
+    int l=(int)strlen( s ) * (int)sizeof(wchar_t);
     if ( l >= MaximumLength )
       init( l );
     MultiByteToWideChar( CP_ACP, 0, s, l/sizeof(wchar_t), Buffer, l );
@@ -70,7 +70,7 @@ public:
   }
   void Set( wchar_t *s )
   {
-    int l=wcslen( s ) * sizeof(wchar_t);
+    int l=(int)wcslen( s ) * (int)sizeof(wchar_t);
 	  if ( l >= MaximumLength )
 		  init( l );
 	  memmove(Buffer,s,l);
